+++ NEWS.md
@@ -6,6 +6,13 @@
   called `rouver.original_path_info` that contains the original
   `PATH_INFO` value.
 
+# News in rouver 2.1.1
+
+## Improvements
+
+* `rouver.test`: Use build-in `assert` statements instead of assertions
+  from the `asserts` package.
+
 # News in rouver 2.1.0
 
 ## API Additions